Rename following public/private API convention
