+++ src/python/zquantum/qaoa/estimators/__init__.py
@@ -1,2 +1,2 @@
-from .cvar_estimator import CvarEstimator
-from .gibbs_estimator import GibbsObjectiveEstimator
+from ._cvar_estimator import CvarEstimator
+from ._gibbs_estimator import GibbsObjectiveEstimator
